ref: Move all tests into one file
